moved components to diff folder strucutre. view past commits for set up. test 1.4
