+++ frontend/src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import Login from "./login";
+import Login from "./components/login";
 
 export default function Home() {
